# Conflicts:
#	frontend/components/ChatComponent.tsx
#	goldenverba/components/schema/schema_generation.py

+++ goldenverba/components/reader/pdfreader.py
@@ -1,17 +1,17 @@
-import glob
 import base64
+import glob
 import os
+from datetime import datetime
+from pathlib import Path
 
 from wasabi import msg
-from pathlib import Path
-from datetime import datetime
 
-from goldenverba.components.reader.interface import Reader, InputForm
 from goldenverba.components.reader.document import Document
+from goldenverba.components.reader.interface import InputForm, Reader
 
 try:
     from PyPDF2 import PdfReader
-except Exception as e:
+except Exception:
     msg.warn("PyPDF2 not installed, your base installation might be corrupted.")
 
 
@@ -30,10 +30,10 @@ class PDFReader(Reader):
 
     def load(
         self,
-        bytes: list[str] = [],
-        contents: list[str] = [],
-        paths: list[str] = [],
-        fileNames: list[str] = [],
+        bytes: list[str] = None,
+        contents: list[str] = None,
+        paths: list[str] = None,
+        fileNames: list[str] = None,
         document_type: str = "Documentation",
     ) -> list[Document]:
         """Ingest data into Weaviate
@@ -42,9 +42,16 @@ class PDFReader(Reader):
         @parameter: paths : list[str] - List of paths to files
         @parameter: fileNames : list[str] - List of file names
         @parameter: document_type : str - Document type
-        @returns list[Document] - Lists of documents
+        @returns list[Document] - Lists of documents.
         """
-
+        if fileNames is None:
+            fileNames = []
+        if paths is None:
+            paths = []
+        if contents is None:
+            contents = []
+        if bytes is None:
+            bytes = []
         documents = []
 
         # If paths exist
@@ -61,28 +68,26 @@ class PDFReader(Reader):
                         msg.warn(f"Path {data_path} does not exist")
 
         # If bytes exist
-        if len(bytes) > 0:
-            if len(bytes) == len(fileNames):
-                for byte, fileName in zip(bytes, fileNames):
-                    decoded_bytes = base64.b64decode(byte)
-                    with open(f"{fileName}", "wb") as file:
-                        file.write(decoded_bytes)
+        if len(bytes) > 0 and len(bytes) == len(fileNames):
+            for byte, fileName in zip(bytes, fileNames):
+                decoded_bytes = base64.b64decode(byte)
+                with open(f"{fileName}", "wb") as file:
+                    file.write(decoded_bytes)
 
-                    documents += self.load_file(f"{fileName}", document_type)
-                    os.remove(f"{fileName}")
+                documents += self.load_file(f"{fileName}", document_type)
+                os.remove(f"{fileName}")
 
         # If content exist
-        if len(contents) > 0:
-            if len(contents) == len(fileNames):
-                for content, fileName in zip(contents, fileNames):
-                    document = Document(
-                        name=fileName,
-                        text=content,
-                        type=document_type,
-                        timestamp=str(datetime.now().strftime("%Y-%m-%d %H:%M:%S")),
-                        reader=self.name,
-                    )
-                    documents.append(document)
+        if len(contents) > 0 and len(contents) == len(fileNames):
+            for content, fileName in zip(contents, fileNames):
+                document = Document(
+                    name=fileName,
+                    text=content,
+                    type=document_type,
+                    timestamp=str(datetime.now().strftime("%Y-%m-%d %H:%M:%S")),
+                    reader=self.name,
+                )
+                documents.append(document)
 
         msg.good(f"Loaded {len(documents)} documents")
         return documents
@@ -91,7 +96,7 @@ class PDFReader(Reader):
         """Loads .pdf file
         @param file_path : Path - Path to file
         @param document_type : str - Document Type
-        @returns list[Document] - Lists of documents
+        @returns list[Document] - Lists of documents.
         """
         documents = []
         full_text = ""
@@ -133,7 +138,7 @@ class PDFReader(Reader):
             # Loop through each file
             for file in files:
                 msg.info(f"Reading {str(file)}")
-                with open(file, "r", encoding="utf-8") as f:
+                with open(file, encoding="utf-8"):
                     documents += self.load_file(file, document_type=document_type)
 
         msg.good(f"Loaded {len(documents)} documents")
